Handle function calling logic in argparse instead of using equals operator on strings. Also add aliases for all commands and move some functions under refs/ to reduce code file size
